+++ R/aaa.R
@@ -2,7 +2,7 @@
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(
     'Welcome to NAIR: Network Analysis of Immune Repertoire.\n',
-    'Get started using `vignette("NAIR", package = "NAIR")`, or by visiting\n',
+    'Get started using `vignette("NAIR")`, or by visiting\n',
     'https://mlizhangx.github.io/Network-Analysis-for-Repertoire-Sequencing-/'
   )
 }